requirements.txt
RPi.GPIO>=0.7.0
Pillow>=9.0.0
smbus2>=0.4.1
bleak>=0.19.0
numpy>=1.21.0
adafruit-circuitpython-max30102>=1.1.0
python-dotenv>=0.19.0
requests>=2.26.0
mpu6050-raspberrypi>=1.2.0
